requirements.txt
flask==2.0.1
crewai
google-search-results
python-dotenv
tavily-python 